requirements.txt
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
asttokens==3.0.0
autopep8==2.3.1
avro-python3==1.10.2
azure-core==1.32.0
azure-functions==1.21.3
azure-identity==1.19.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.9.0
azure-keyvault-keys==4.10.0
azure-keyvault-secrets==4.9.0
azure-monitor-opentelemetry-exporter==1.0.0b32
azure-monitor-query==1.4.0
azure-storage-blob==12.24.0
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
cryptography==44.0.0
debugpy==1.8.11
decorator==5.1.1
dependency-injector==4.44.0
Deprecated==1.2.15
distro==1.9.0
executing==2.1.0
fastapi==0.115.6
fixedint==0.1.6
flake8==7.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.31.0
isodate==0.7.2
jedi==0.19.2
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mccabe==0.7.0
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.1
oauthlib==3.2.2
openai==1.58.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.48
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.1
pyodbc==5.2.0
pyspark==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
wrapt==1.17.0
zipp==3.21.0
